Moving all classes from `storage` directly to `internal`
